try fix prod errors

+++ api/src/modules/picture/picture.service.ts
@@ -53,7 +53,11 @@ export class PictureService {
 
   async remove(id: number): Promise<void> {
     const picture = await this.prisma.picture.findUnique({ where: { id } });
-    fs.rmSync(`/pictures/${picture.url}`);
+
+    if (fs.existsSync(`/pictures/${picture.url}`)) {
+      fs.rmSync(`/pictures/${picture.url}`);
+    }
+
     await this.prisma.picture.delete({ where: { id } });
   }
 }
